ci: fix format

+++ script/AnvilInit.s.sol
@@ -27,10 +27,7 @@ contract AnvilInitScript is Script {
         uint256 runnerPrivateKey = vm.envUint("SCRIPT_RUNNER_PRIVATE_KEY");
         address runnerAddress = vm.addr(runnerPrivateKey);
         uint256 nonce = getNonce(runnerAddress);
-        require(
-            nonce == 0,
-            "Init runner nonce should be 0, aborting deployment."
-        );
+        require(nonce == 0, "Init runner nonce should be 0, aborting deployment.");
         require(runnerAddress == mustRunnerAddress, "Unmatched deployer");
 
         vm.startBroadcast(runnerPrivateKey);
@@ -40,24 +37,17 @@ contract AnvilInitScript is Script {
 
         // deploy Counter
         bytes memory initCode = abi.encodePacked(type(Counter).creationCode);
-        address computedCounterAddress = create2Deployer.computeAddress(
-            salt,
-            keccak256(initCode)
-        );
+        address computedCounterAddress = create2Deployer.computeAddress(salt, keccak256(initCode));
         address deployedAddress = create2Deployer.deploy(salt, initCode);
         console.log("Counter contract address: ", deployedAddress);
-        require(
-            computedCounterAddress == deployedAddress,
-            "create2 counter computed-address invalid"
-        );
+        require(computedCounterAddress == deployedAddress, "create2 counter computed-address invalid");
         Counter(deployedAddress).increment();
-        uint num = Counter(deployedAddress).number();
+        uint256 num = Counter(deployedAddress).number();
         require(num == 1, "counter number != 1");
 
         // deploy test20 by create2
         initCode = abi.encodePacked(
-            type(Test20).creationCode,
-            abi.encode("Local Test20", "Test20", uint8(6), uint256(10000000000))
+            type(Test20).creationCode, abi.encode("Local Test20", "Test20", uint8(6), uint256(10000000000))
         );
         address test20Address = create2Deployer.deploy(salt, initCode);
         console.log("Test20 contract address: ", test20Address);
@@ -70,16 +60,11 @@ contract AnvilInitScript is Script {
         // console.log("test20: create2 deployer blance of test20: ", bal);
 
         // deploy nft by create2
-        initCode = abi.encodePacked(
-            type(NFT).creationCode,
-            abi.encode("Hero NFT", "Hero", "blank://hero-url")
-        );
+        initCode = abi.encodePacked(type(NFT).creationCode, abi.encode("Hero NFT", "Hero", "blank://hero-url"));
         address nftAddress = create2Deployer.deploy(salt, initCode);
         console.log("Hero NFT contract address: ", nftAddress);
         uint256 mintValue = NFT(nftAddress).MINT_PRICE();
-        uint256 token_id = NFT(nftAddress).mintTo{value: mintValue}(
-            runnerAddress
-        );
+        uint256 token_id = NFT(nftAddress).mintTo{value: mintValue}(runnerAddress);
         require(token_id == 1, "init mint hero token != 1");
 
         vm.stopBroadcast();
